// services/digimal/api.ts
export interface GetFilterItemsAndProductsByUserResult {
    /**
     * 
     * @type {GetProductsFilterByUserResultInfo}
     * @memberof GetFilterItemsAndProductsByUserResult
     */
    'productsList'?: GetProductsFilterByUserResultInfo;
    /**
     * 
     * @type {GetFilterItemsByUserResultInfo}
     * @memberof GetFilterItemsAndProductsByUserResult
     */
    'filters'?: GetFilterItemsByUserResultInfo;
    /**
     * 
     * @type {Array<GetProductSortResult>}
     * @memberof GetFilterItemsAndProductsByUserResult
     */
    'sorts'?: Array<GetProductSortResult> | null;
}
/**
 * 
 * @export
 * @interface GetFilterItemsByUserResultInfo
 */
export interface GetFilterItemsByUserResultInfo {
    /**
     * 
     * @type {Array<GetProductTypeFilterByUserResult>}
     * @memberof GetFilterItemsByUserResultInfo
     */
    'productTypes'?: Array<GetProductTypeFilterByUserResult> | null;
    /**
     * 
     * @type {Array<GetProductBrandFilterByUserResult>}
     * @memberof GetFilterItemsByUserResultInfo
     */
    'brands'?: Array<GetProductBrandFilterByUserResult> | null;
    /**
     * 
     * @type {Array<GetProductPropertyFilterByUserResult>}
     * @memberof GetFilterItemsByUserResultInfo
     */
    'properties'?: Array<GetProductPropertyFilterByUserResult> | null;
    /**
     * 
     * @type {Array<GetProductSpecificationFilterByUserResult>}
     * @memberof GetFilterItemsByUserResultInfo
     */
    'specifications'?: Array<GetProductSpecificationFilterByUserResult> | null;
    /**
     * 
     * @type {GetProductsPriceRangeResult}
     * @memberof GetFilterItemsByUserResultInfo
     */
    'productPriceRange'?: GetProductsPriceRangeResult;
}
// ...
export interface GetProductBrandFilterByUserResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductBrandFilterByUserResult
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductBrandFilterByUserResult
     */
    'persianName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductBrandFilterByUserResult
     */
    'englishName'?: string | null;
}
// ...
export interface GetProductFilterByUserResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductFilterByUserResult
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'persianName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'englishName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'introduction'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'expertCheck'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductFilterByUserResult
     */
    'brandId'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'brandPersianName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'brandEnglishName'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductFilterByUserResult
     */
    'b2CPrice'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetProductFilterByUserResult
     */
    'remainingCount'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'barCode'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductFilterByUserResult
     */
    'productTypeId'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'productPersianTypesName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductFilterByUserResult
     */
    'productEnglishTypesName'?: string | null;
    /**
     * 
     * @type {Array<GetProductMediaResult>}
     * @memberof GetProductFilterByUserResult
     */
    'mediaList'?: Array<GetProductMediaResult> | null;
    /**
     * 
     * @type {GetSpecialSaleFilterByUserResult}
     * @memberof GetProductFilterByUserResult
     */
    'specialSale'?: GetSpecialSaleFilterByUserResult;
}
// ...
export interface GetProductPropertyFilterByUserResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductPropertyFilterByUserResult
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductPropertyFilterByUserResult
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductPropertyFilterByUserResult
     */
    'value'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof GetProductPropertyFilterByUserResult
     */
    'isFilter'?: boolean;
}
// ...
export interface GetProductSortResult {
    /**
     * 
     * @type {string}
     * @memberof GetProductSortResult
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductSortResult
     */
    'persianTitle'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductSortResult
     */
    'value'?: number;
}
/**
 * 
 * @export
 * @interface GetProductSpecificationFilterByUserResult
 */
export interface GetProductSpecificationFilterByUserResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductSpecificationFilterByUserResult
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductSpecificationFilterByUserResult
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductSpecificationFilterByUserResult
     */
    'value'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof GetProductSpecificationFilterByUserResult
     */
    'isFilter'?: boolean;
}
// ...
export interface GetProductTypeFilterByUserResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductTypeFilterByUserResult
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetProductTypeFilterByUserResult
     */
    'parentId'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductTypeFilterByUserResult
     */
    'persianName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductTypeFilterByUserResult
     */
    'englishName'?: string | null;
}
// ...
export interface GetProductsFilterByUserResultInfo {
    /**
     * 
     * @type {Array<GetProductFilterByUserResult>}
     * @memberof GetProductsFilterByUserResultInfo
     */
    'productsInfo'?: Array<GetProductFilterByUserResult> | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductsFilterByUserResultInfo
     */
    'totalCount'?: number;
}
/**
 * 
 * @export
 * @interface GetProductsPriceRangeResult
 */
export interface GetProductsPriceRangeResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductsPriceRangeResult
     */
    'minPrice'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetProductsPriceRangeResult
     */
    'maxPrice'?: number;
}